Allow CPDF_CrossRefTable::AddNormal() to overwrite free entries
AddNormal() currently only checks the generation number in the cross
reference table to decide if a new entry can overwrite or not. When free
entries have a generation number of 65535, AddNormal() never allows free
entries to be overwritten, as normal entries typically have a generation
number that is close to 0 in valid.

Fix the generation number check to only the generation number for
non-free entries. Now PDFium will correctly combine cross reference
tables and streams for certain PDFs with appended sections that contain
this property.

Bug: 345274934
Change-Id: I98fad76a7bd684419798f3a023357d40001c196e