Add a pixel test for yet another edited PDF that regressed
The test PDF is yet another variation of rectangles.pdf. For this file,
the top part is the original PDF. The original PDF has a cross reference
stream instead of a cross reference table. It also has the content
stream set to object 3 0, but that object is a free entry. Thus the
original PDF renders blank. Next, the appended section adds a normal
object 3 0 with the content stream from rectangles.pdf. So all together,
the PDF should render just like rectangles.pdf. However, the PDF parser
fails to overwrite the free entry in the original PDF with the normal
entry from the appended section, so the PDF currently renders blank. As
such, the pixel test starts out suppressed.

Bug: 345274934
Change-Id: I91368171594b05d13cb4cc3cc96b67b965830201