Do not hard code the generation number in CPDF_CrossRefTable::SetFree()
CPDF_CrossRefTable::SetFree() currently hard codes the generation number
for free objects to 65535, which is not correct. Instead, make the
callers pass it in:

- CPDF_Parser::MergeCrossRefObjectsData() already know the generation
  number.
- CPDF_Parser::ProcessCrossRefStreamEntry() can read it out of the
  cross-reference entry.

With the generation number no longer hard coded to the maximum possible
value, CPDF_CrossRefTable::AddNormal() can correctly overwrite free
entries. As a result, bug_345274934.pdf renders correctly and its pixel
test results can be unsuppressed, but only with the addition of a
missing expectations file.

Bug: 345274934
Change-Id: I98fad76a7bd684419798f3a023357d40001c196e