Filter out bad font base encodings when flattening
FPDFPage_Flatten() draws annotations into the content stream. From the
perspective of a PDF reader that only reads the content stream, the
flattened PDF potentially uses more fonts than the original PDF. If the
PDF reader is strict about validating fonts, it can then detect
malformed fonts and reject them.

Avoid a case of this by removing font encodings that are out-of-spec
during the flattening process. Then PDF readers that are strict about
validating font encodings can fallback to default values instead of
choking on the bad values.

Update FPDFFlattenEmbedderTest.FlatWithBadFont to reflect this change
and fulfill the TODO related to this issue.

Bug: 344775293
Change-Id: I1d01dd233caffcaecd6a9d1f391d1bf56da075bb