Spanify input side of LZWDecompressor.
Use a span in place of a pointer/length pair. Interestingly, this now
avoids a situation where an error return previously could occur between
updating the pointer and decrementing the size. This avoids some
UNSAFE_BUFFER usage but some still remains on the output side of
the decompressor.

There is a small change to the return value of Unfishished whenever
the span is empty regardless of whether one had ever been supplied.
Previously, this was considered an error. This is not expected to
affect callers outside of tests.

Bug: 42271176
Change-Id: I48cc36b3ef14b68ac5ca7277fc0a5280d79e4df3