Add a pixel test for generation numbers in cross-reference streams
Demonstrate yet another edited PDF that fails to parse correctly with
bug_349972030.pdf. Here, the initial cross-reference stream has a free
object with a generation number of 1. The appended cross-reference
stream tries to reuse the same object number with a generation number
of 2. The PDF should render the same as rectangles.pdf, but currently
does not because the generation numbers are not parsed correctly. So its
test result is suppressed.

Bug: 349972030
Change-Id: I69dcc13252771cb07d0df9cb79d2a2b9ee05e6b5