Remove CFX_DIBitmap::SetUniformOpaqueAlpha() return value
SetUniformOpaqueAlpha() always succeeds, because the only condition that
can cause it to fail never happens in its 2 callers. Change the
conditional that causes SetUniformOpaqueAlpha() to fail into a CHECK().

This is partly because https://pdfium-review.googlesource.com/115412
forced callers that creates CFX_DIBitmaps to make sure the Create() call
succeeded.

In CGdiDisplayDriver::GetDIBits(), move the SetUniformOpaqueAlpha() call
into the only branch where it can happen.

Change-Id: Ib41ad6ece695d54b2cab493b1b46e1b8c713419b