Check if bitmaps are premultiplied in CFXDIBitmapTest cases
Add many CFXDIBitmap::IsPremultiplied() calls to CFXDIBitmapTest cases.
Demonstrate how the "cleared" bitmap is in a confused state where
IsPremultiplied() returns false, yet UnPreMultiply() is not a no-op.

Make IsPremultiplied() public in CFX_DIBBase and CFXDIBitmap so the test
cases do not have to be friends.

Also update bug numbers for this issue to the Buganizer bug number.

Bug: 42271020
Change-Id: I0406c4186ac0c9b208133fed4f9bffbd5a6e3e6d