Add EmbedderTest::LoadScopedPage()
Implement a LoadPage() variant that returns a ScopedEmbedderTestPage.
This scoper automatically calls EmbedderTest::UnloadPage() when it goes
out of scope. Convert a few FPDFViewEmbedderTest test cases to call
LoadScopedPage() instead of LoadPage() and get rid of the UnloadPage()
calls.

This also makes tests more robust if they return early due to gtest
assertion failures. The UnloadPage() will still run and the test won't
hit a CHECK() due to the unbalanced LoadPage() / UnloadPage() counts.

Change-Id: I45770f06777d18bd31c5e4e772b69caf2d3e8515