Limit bitmap formats in ProgressiveDecoder::StartDecode()
Based on all the callers, limit the bitmap formats StartDecode() accepts
to just 3. Then delete all the code in ProgressiveDecoder that assume
the other formats can be used.

Along the way, get rid of some method parameters that are always the
same member variables, and just use those member variables directly.
Rename GetTransMethod() to SetTransMethod() as well, since it does not
return anything.

Change-Id: I3fe72a8858e1451cf1d61323c80a6c1e7feb0170