Remove dead code in ProgressiveDecoder::HorzTable::CalculateWeights()
The two arguments to CalculateWeights() always have the same value, so
remove one of the parameters and rename the remaining one from "len" to
"width". Then the `scale` within CalculateWeights() is always 1, and the
code to handle scaling can be removed. Along the way, minimize the scope
of the UNSAFE_TODO() inside CalculateWeights().

Change-Id: Id68c5186d073f782677cd8c36b763823ac4503ad