Add PreMultiplyColor()
Add a Skia-only utility function to premultiply FX_ABGR_STRUCT and
similar structs. This will be used in CFX_DIBitmap::Clear(), where the
color parameter is always going to be specified with straight alpha. It
may be used in other places as well, since PDFium internally specifies
FX_ARGB values with straight alpha.

Bug: 42271020
Change-Id: I2fbf694495d78e79c86035c526cbb0ee9630c77f