Limit formats supported by CFX_DIBBase::ConvertTo()
There are only a few callers to CFX_DIBBase::ConvertTo(), and they only
pass in one of two possible bitmap formats. Assert this pre-condition in
ConvertTo(), and delete code that assumes otherwise.

This removes one CFX_DIBitmap::SetUniformOpaqueAlpha() call, and the
remaining call is Windows-only, so mark SetUniformOpaqueAlpha() as such.

Change-Id: I155f08f225ba9775814e88e0271ba5a6fa04136a