Delete ConvertBuffer_1bppPlt2Gray()
There are no 1 BPP bitmap to `FXDIB_Format::k8BppMask` conversions in
the code base, except in CFX_DIBBase::ConvertBuffer() itself. Since the
caller is in a code path where the source bitmap does not have palettes,
the path for 1 BPP palette conversion can be removed. Replace it with a
CHECK() to make sure there is no palette. Also change ConvertBuffer() to
call the helper function directory, instead of calling itself.

Change-Id: If8cdccd231499f8fe725fc3aec2055f6b717a616