Add special FPDFBitmap_FillRect() handling for 24 and 32 BPP bitmaps
If Skia were to truly use pre-multiplied bitmaps internally, then the
current FPDFBitmap_FillRect() implementation would take an
un-premultiplied bitmap, pre-multiply it, let Skia fill it, and
un-premultiply it. Avoid this extra work by changing
FPDFBitmap_FillRect() to write to 32 BPP bitmaps directly.

Similarly, handle filling 24 BPP bitmaps directly in
FPDFBitmap_FillRect() to avoid a roundtrip conversion to 32 BPP and
back.

This also improves the rendering in an embedder test that uses
FPDFBitmap_FillRect(). There, the current Skia-specific test
expectations are for an image that is a drawn a bit too dark.

Bug: 42271020
Change-Id: Id156780f3ea4aa7479dc933b2da5ca192b2e27a1