Check `FXDIB_Format::kArgbPremul` in CFX_DIBBase::IsPremultiplied()
Remove `CFX_DIBitmap::m_nFormat` and use `CFX_DIBBase::m_Format` as the
canonical source of bitmap format information. Also remove
CFX_DIBitmap::ForcePreMultiply(), as bitmaps can be pre-multiplied to
start with if they are of `FXDIB_Format::kArgbPremul`.

Update some test expectations to match this change. Most of the changes
are off by 1 pixel values due to rounding errors. When comparing the
old and new bitmaps, they are not pixel perfect according to
pdfium_diff, but will pass with --fuzzy.

Bug: 42271020
Change-Id: I7726b3915602c0ca1845a19e28f9d2d6cc107faf