Fix UNSAFE_TODOs in RgbByteOrderTransferBitmap()
Switch to accessing bitmaps one scanline at a time, using fxcrt::Zip()
to iterate through the pixels. This fixes a crash that happens when
iterating through the entire buffer and trying to calculate the scan
lines manually.

Along the way, change the cheap DCHECKs into CHECKs, and add a CHECK()
to make ensure the bitmap is `FXDIB::kRgb` before reinterpreting the
scanline.

Bug: 42271147,42271176
Change-Id: I7fab5acebf9ca3e33e845beb99077a9cb37c5066