Use third_party/fast_float to convert strings to floats/doubles instead
The existing implementation of StringTo() rounds to the incorrect
float/double value for values with precision error.

Replace the implementation to use third_party/fast_float, which fixes
the rounding error, is easier to read, and is supposedly fast.

Update pixel tests and checksums. There should be little to no visible
difference in the PDFs.

Bug: 364338789, 366309453
Change-Id: I51f4aa9b7fa7a53306d8a2bd604547cfb079eb30