Creating a util function to compare two FS_RECTF objects
This util function aims to ease the process of comparing two FS_RECTF
objects.
The process is currently carried out through a chain of ASSERT_EQ
function calls for each of the fields(top,left,bottom,right) of the
FS_RECTF struct.
I could not find the appropriate directory to place the code in the chrome build so I went ahead and made the change in the PDFIUM build.Hope that works out.

Bug: 42270440
Change-Id: Ied7467758817efa8f9a1ce0e5f767a462918cb22