Creating a util function to compare two FS_RECTF objects
This util function aims to ease the process of comparing two FS_RECTF
objects.
The process is currently carried out through a chain of ASSERT_EQ
function calls for each of the fields(top,left,bottom,right) of the
FS_RECTF struct.

Pathchset 2-3:
Changed the implementation with the order of the fields in the same order as the FS_RECTF struct definition.
Also used the function call in a test to prove it's working.

Patchset 4-5:
Corrected the BUILD.gn files to include the new helper functions in the embedder_test targets instead of pdfium_test.
C++ coding style implemented.
Changed the function signature to match parameter names in the  ASSERT_FLOAT_EQ() definition (in the .cc file) .

Patchset-6:
Corrected the function signature in the .h file as well.

Patchset 7-8:
Merge conflicts fixed.

Bug: 42270440
Change-Id: Ied7467758817efa8f9a1ce0e5f767a462918cb22