Added a util function to compare two FS_RECTF objects.
Uses a chain of ASSERT_FLOAT_EQ() calls for each of the fields in the struct.
Comparison carried out in the same order as the definition of fields.
Updated the BUILD.gn file to include this helper function in the embedder_tests.
Error made during conflict resolution was fixed.
Removed the function check_unsigned_shorts() mistakenly added during conflict resolution.

Bug: 42270440
Change-Id: Ied7467758817efa8f9a1ce0e5f767a462918cb22