Added a util function to compare two FS_RECTF objects.
This process is currently carried out using ASSERT_EQ() function
calls for each of the fields in the struct.
Instead this will now be carried out by single function call to
CompareFS_RECTF(). It also uses ASSERT_FLOAT_EQ() instead of
ASSERT_EQ() to prevent any unwanted behavior during comparison of
floats. Comparison is carried out in the same order as the 
definition of the fields.
The files where this function resides
(testing/helpers/compare_coordinates.h (/.cc)) have been added to
the embedder_tests target in the BUILD.gn file.
The utility function has been used in the
FPDFTextEmbedderTest.CroppedText test to demonstrate it's working.

Bug: 42270440
Change-Id: Ied7467758817efa8f9a1ce0e5f767a462918cb22