Limit the CID code values in CPDF_ToUnicodeMap
It is not perfectly clear in the Adobe CMap and CIDFont Files
Specification, but all the examples and some auxiliary docs indicate
that the CID code values are never above 65535. Therefore, check for
that in CPDF_ToUnicodeMap, and stop worrying that code ranges can have
integer overflows.

Also check for invalid ranges where the low value is higher than the
high value. Adjust results for existing integer overflow cases. Expand
testing to cover new checks.

Bug: 41489089
Change-Id: I4c1e0f1a20ed239c925c75fa6f98863c46f07b7a