Spanify FXSYS_ToUTF16BE()
Pass a fixed-size span of the required size into FXSYS_ToUTF16BE()
instead of a pointer of unknown size. Also change the return type to be
a span that represents the used portion of the input span.

Along the way, prevent a copy in fpdf_edittext.cpp by using std::move().

Bug: 42270715
Change-Id: I31897efa547f8a2c07f9928786d6fa90cbe652b2