Fix crash in CPDF_InteractiveForm::AddTerminalField()
AddTerminalField() attempts to sanitize its in-memory representation by
making adjustments to the field dictionary. One such entry in the field
dictionary is the name (/T) entry. When the /T dictionary value is a
reference, AddTerminalField() clones it to a non-reference object,
possibly so the object can be modified later without affecting other
parts of the PDF that references the same object. In doing so,
AddTerminalField() makes 2 mistakes:

1) It does not check the cloned object's type. If the cloned object is a
   stream, it cannot be stored in the dictionary. Fix this by making
   sure the object is a string, as non-string /T values are garbage
   values anyway.
2) In the fallback path when cloning fails, it stores an empty name as
   the /T value. The value should be a string instead.

Fix these two issues and do some style / variable name cleanup along the
way.

Bug: 379844656
Change-Id: If1afda6f49e34a32bf9f512a5ba6c5caec0c7313