Check for out of range values in CPDF_Font::GetFontWeight()
The font weight is inferred from the /StemV value. If the value is wonky
in some way, currently the code only checks for integer overflows and
falls back to the normal weight in that case. Add more checks for weight
values that are outside of the expected range and do the same fallback.

Bug: 380018383
Change-Id: Ia982f99ffdc60678c3b185fbe1a0d1024f0af243