Assume no trailing space for CID font characters if it is not specified.
When default width value is not provided for a CID font ("DW" in font
dictionary), it is assumed to be 1000, and if a character's width is not
explicitly provided ("W" array in font dictionary), default width is
used in computation of the charater width with trailing space
(`GetCharWidthF`).

Assume trailing space to be zero if it is not explicitly provided either
as character width or default character width.

Bug: 376781381
Change-Id: Iccbbebc93beff4ab082ac3f98f671bc9b69363bc