Demonstrate page size change after flattening
Add a test which demonstrates some undesireable behavior for a PDF
which incorrectly has the MediaBox in the page tree node instead of the
page object.  Viewers are able to correctly display the PDF even though
the MediaBox is misplaced.  Once the PDF is changed with an annotation
markup, flattened, and saved, the resulting PDF has a MediaBox entry in
the correct place.  However, the dimensions in the MediaBox are not
consistent with the original choice that is shown when viewing the
original PDF, and is a mismatch to the saved page dimensions.

Bug: 387312214
Change-Id: Ifa26fb99987a2a95376c315bf9ef1f16b3990d41