Fix unnecessary tree traversal in SearchNameNodeByNameInternal()
When SearchNameNodeByNameInternal() is searching the name tree for
`csName`, it encounters nodes with limits, designating a range of what
possible values can be found in its subtree. Logically, when one is
searching for `csName`, and encounters a node where `csName` is greater
than the node's greatest limit, that subtree should be skipped, and the
next sibling node (if it exists) should be processed instead. This was
the behavior prior to https://pdfium-review.googlesource.com/8271, but
that CL introduced a regression where the subtree still gets traversed
anyway. This behavior might make sense when trying to add to the name
tree, where there needs to be some tree balancing, but it is unnecessary
when only searching the tree, leading to slow performance for PDFs with
large name trees.

Fix this behavior by skipping the subtree when only trying to search for
`csName`.

Fixed: 372523840
Change-Id: I805e0a32fbd6ffdbece6f831f1be6d3a977587fc