Fix JPEG2000 image decoding with an indexed colorspace
Fix a regression in JPEG2000 image decoding, where the PDF specifies an
indexed colorspace for a grayscale JPEG2000 image. Add a test case for
this, where the JPEG2000 started as a PNG with a 2-bit colormap. The PNG
then gets converted to JPEG2000 using Imagemagick as follows:

convert input.png -colorspace Gray output.jp2

Bug: 393395940
Change-Id: I1a0265091a5ce906d3461fb803984c7345f6fe8e