fixup_pdf_template: Don't count final newline in {{streamlen}}
Expansion of `{{streamlen}}` used to include the newline in front of
`endstream`.

macOS's Preview.app cannot load JPEG2000 files if they have an
additional `\n` byte at the end of the data, so don't count this
final newline.

With this, macOS's Preview.app now displays the PDF files generated
by fixup_pdf_template.py much better for
testing/resources/pixel/jpxdecode*.in

This requires adding newlines in a couple of tests that depend on
stream contents actually ending in a newline:

* bug_1236805: Apparently the type3 charprocs parser doesn't treat
               EOF as a token separator (?)
* bug_1395648: Added in https://pdfium-review.googlesource.com/102930
               as a regression test with a missing `/Filter` line.
               In the previous image in the text, the `ASCIIHexDecode`
               decode filter stripped newlines, but here it doesn't and
               the test needs the newline.

Bug: 394805680
Change-Id: I91610fe7d85080e5d4e0b3c52e2d5d10ff4b985d