Demonstrate FPDFAnnot_GetColor() does not work with normal page loading
Change 2 test cases that call FPDFAnnot_GetColor() to use
EmbedderTest::LoadScopedPage() like most other test cases, instead of
LoadPageNoEvents(). This causes FPDFAnnot_GetColor() to fail, and thus
demonstrating the bug. Update the test expectations to expect the
failure, and temporarily disable the code that checks the color values.

Also recognize that LoadScopedPage() triggers PDFium form loading, and
that can change the rectangles for some annotations. Since there is no
good way to recover the original rectangles, make a note of this case in
the FPDFAnnot_GetRect() documentation and adjust affected test results.

Bug: 396056603
Change-Id: Ib0352aa5b789dec3ceb9faa16cce30c67922d85c