Clamp PDFium indexed color space hival
The PDF spec indicates that the high value for an indexed color space
must be in the range of 0 to 255.  Should a color space violate this,
failing out of loading the color space results in images which rely
on the color space to not display.

Modify the approach to instead clamp the specified high value to be
within the spec range limit.  This slight corrective action can
sometimes allow an image to be displayed.  This allows the behavior
to match some other PDF readers.

Fixed: 372749732
Change-Id: Ib77d46d04e9da6dbc0e9345653c38241a15dc45e