Fix rendering for freetext_annotation_without_da.in
Change GenerateFreeTextAP() to initialize the /AcroForm dictionary if it
does not exist. Reuse InitDict() from cpdf_interactiveform.cpp to do
this by exposing it publicly. To make InitDict() work as expected and
create a reasonable default appearance entry:

- Change its font selection logic to avoid overwriting the "standard"
  font with the "native" font, if there is no "native" font.
- Use a font size of 12 instead of 0.

Also update DEPS with new corpus test expectations to match this change:

https://pdfium.googlesource.com/pdfium_tests/+log/0d854e64bed7..039b875bf673

Bug: 380434959
Change-Id: I7131488d8f184bfb1bd4fc698a048fd9781d99f6