Change CPDF_PageContentGenerator to regenerate all stream
In CPDF_PageContentGenerator, instead of only regenerating the modified
streams, regenerate all stream. While this is more work, it should fix
incorrect output in cases where a page object is defined in more than 1
stream. For instance:

1) Stream content loss

Stream A contains an image and part of a path.
Stream B contains the rest of the path.
When the image is modified, if only Stream A gets regenerated, then the
new Stream A only contains the image and the part of the path is lost.

2) Leftover stream content

Stream A contains an image and part of a path.
Stream B contains the rest of the path.
When the path is modified, if only Stream B gets regenerated, then the
part of the path that still exists in Stream A is going to cause
confusion.

As a result of this change:

- CPDF_PageObjectHolder no longer needs to track dirty streams.
- WriteColorToStream() needs to work for all colors, not just gray and
  RGB. Then CPDF_Color::IsColorSpaceRGB() and
  CPDF_Color::IsColorSpaceGray() can be deleted because their only
  caller is gone.
- CPDF_Color::GetRGB() needs to handle null `cs_` value.
- A second PDF generation bug gets fixed.

Bug: 42270554, 405433817
Change-Id: Ie71c42a19b13759355f5de023e5a16a9dd974e49