Let FPDFPageObj_SetMatrix() reset dirty bit if restoring image matrix
With existing public image APIs, embedders may want to first call
FPDFPageObj_SetMatrix() to set an image to a desired size, render, and
call FPDFPageObj_SetMatrix() to restore the original matrix. This has
the side effect of marking the image object as being dirty, which can
exacerbate issues in CPDF_PageContentGenerator.

Avoid this by remembering a page object's initial matrix, and have a
second dirty bit to track if the matrix has changed, separate from
other modifications. Do this only for image objects for now.

Bug: 405433817, 406540676, 407950121
Change-Id: I0e66ef4e590cd7bcc79c21f31eae2e02724c422f