Improve shared resources dict detection in CPDF_PageContentGenerator
The existing code in CPDF_PageContentGenerator to detect shared page
resources dictionary can miss some cases, like when the resource
dictionary is inlined into the /Pages (not /Page) dictionary. Improve
the code that detects shared resources to look for this case and other
possibly scenarios by checking if the current page's page dictionary and
resources dictionary are in use twice in the same document.

With the additional code, refactor the related code into
IsPageResourceShared().

Bug: 407730667
Change-Id: Ie35100c0fdf7693171f91ec00df094adc36bdce3