Clamp PDFium indexed color space hival
The PDF spec indicates that the high value for an indexed color space
must be in the range of 0 to 255.  Should a color space violate this,
failing out of loading the color space results in images which rely
on the color space to not load, and thus not be displayed by viewers.

Modify the approach to instead clamp the specified high value to be
within the spec range limit.  This slight corrective action can
allow such images to now be loaded.  This makes for a smoother user
experience, and also matches the behavior of some other PDF readers,
such as Firefox and Evince.

Fixed: 372749732
Change-Id: Ib77d46d04e9da6dbc0e9345653c38241a15dc45e