Rework inline image parsing in content streams
Currently, CPDF_StreamContentParser:Handle_BeginImage() will try to
parse inline images in content streams when it sees the "BI" operator.
The parts that deal with the related "ID" and "EI" operators
(Handle_BeginImageData() and Handle_EndImage(), respectively) are
no-ops, since Handle_BeginImage() did all the work.

Rework Handle_BeginImage() and Handle_BeginImageData() so
CPDF_StreamContentParser records their respective keywords's positions.
Then handle the inline image parsing in Handle_EndImage() with those
positions. At this point, CPDF_StreamContentParser know the distance
between the "ID" and "EI" operators, so it knows where the length of the
inline image stream. Pass this length to
CPDF_StreamParser::ReadInlineStream() and prevent it from reading the
"EI" operator or past it.

As a result of this change:
- Handle_BeginImage() no longer needs to look for "ID" and "EI".
- ReadInlineStream() no longer needs to looks for "EI".
- If parsing reaches the end of the stream and there was no "EI", then
  ignore the incomplete inline image. The existing code treated end of
  stream as though it had found "EI".
- bug_1236805.in fails because the code was reading "EI" and past it.
  Fix this by adding a placeholder value so the inline stream is not
  empty.
- The bug_407752631.in test case can be checked in and will no longer
  cause a crash.

To deal with malforms images that may not have "BI", "ID", and "EI"
operators in that order, do some state tracking so that
CPDF_StreamContentParser only looks for "ID" and "EI" when appropriate,
like the existing code.

Bug: 407752631
Change-Id: Ie14908d8ed72c92c2ae880c18f62469a5cf4eef3