Interpolate shadings in correct color space
...and evaluate functions per-pixel.
(If this is a performance issue, build an 1-d LUT and
get colors from there, similar to GetShadingSteps().)

This fixes two bugs:

1. We interpolated gouraud and patch colors in RGB,
   instead of in the shading's color space as dictated
   by the spec. (Especially visible for e.g. LAB color
   spaces.) Now we interpolate in the right color space
   and only convert to RGB when painting a primitive.
   (There's no bug filed for this yet.)

2. We evaluated functions too early, causing us to miss
   function values. Now we evaluate functions only when
   painting a primitive. (If this becomes a performance
   problem, we could precompute an 1D LUT using
   GetShadingSteps(), which is used by the radial and
   axial shadings.)

(*: It looks like arrays of 1-element functions never
worked in PDFium; this CL currently doesn't change that.)

CPDF_MeshStream::ReadColor() no longer returns an
RGB value but all components. The rest is mechanical
fallout from this.

Bug: 406662638,42270551,413597066
Change-Id: I05616987674da10db90deb73ae75a95688d264f8