For shading types 4-7, evaluate function per pixel if present
CPDF_MeshStream::ReadColor() now no longer resolves function values
immediately (at the vertex level), but returns them in the red channel
of the returned color (mapped to the range Decode[4]..Decode[5]).

Over in cpdf_rendershading.cpp, DrawGouraud() (for types 4-5) and
PatchDrawer::Draw() (for types 6-7) now use the interpolated red
channel to evaluate funcs.

Evaluating the function per pixel was a ~8x slowdown for types 4-5,
measured by running

    pdfium_test --render-repeats=500 shade-gouraud-free.pdf

On my system, on trunk this took around 0.13s. With naively evaluating
the function per pixel, it took 1.0s instead. So instead of evaluating
the function per pixel, this now computes a 1D texture with 255 samples
of the function, and then looks up the function result per pixel. With
that, it takes ~0.125s, comparable to before. This is a minor quality
loss, but matches what we do for types 1-3 (it in fact uses the
already-existing GetShadingSteps() function to create the texture),
and it's worth it for the perf.

For patch rendering, I measured using

    pdfium_test --render-repeats=50 shade-tensor.pdf

(Note: 10x fewer repetitions.)

That took 1.0s on trunk, with the naive patch, and with the 1D texture.
Our patch rendering is much slower than our gouraud triangle
rendering, and function evaluation is apparently cheap compared to
all the other work we're doing there. I still went with the 1D texture
there too, as the code is a bit nicer, and it's nice to be self-
consistent.

(Perf test iles from here:
https://github.com/SerenityOS/serenity/blob/master/Tests/LibPDF/shade-gouraud-free.pdf
https://github.com/SerenityOS/serenity/blob/master/Tests/LibPDF/shade-tensor.pdf
)

Bug: 406662638
Change-Id: Id70a62ea31afe29c368fa9ab007e40f69985d1be